Fix linker errors and Windows build failures
- src/hazard/hazard_l.c: restore 'int yylnctr,yysynerr;' as a definition
  (commit 54325bb changed it to 'extern', leaving yysynerr undefined at
  link time; yyerror.c and initprz.c both reference it)

- src/hazard/hazrd4.h: remove hardcoded 'extern' from variables in the
  #ifdef HAZRD4 block (commit b142b79 added 'extern' keyword, turning
  definitions into declarations; hval, stepk1..3, ri, rbase, cflag1..4,
  hz4stop, faststop etc. were undefined at link time)

- src/hazard/cput_v.c: fix function pointer type from (char *,long) to
  (char *,int) to match cput_vv.h and the actual cput/cput_l callee sigs

- src/common/xvputd.c, src/hazard/opnfils.c, src/hazpred/opnfils.c:
  add '#include <stdlib.h>' before the _swab() macro on _WIN32; MinGW
  declares _swab() in stdlib.h, not implicitly

+++ src/hazard/cput_v.c
@@ -4,7 +4,7 @@
 #include "cput_vv.h"
 #include "hazard.h"
 
-void cput_v(int dashct,int dashat,void (*cputfn)(char *,long)){
+void cput_v(int dashct,int dashat,void (*cputfn)(char *,int)){
   hzfskp(1);
   hzfxpr('-',dashct,dashat);
   hzfskp(1);
